[vim] change leader to space to match kickstart
I am trying to reduce the customness in my vim life
Everything was weirdly self rolled so far.
And yes thats also what vim is for.

But I am thinking to follow what ever kickstart recommends.

They use space as leader.

https://github.com/nvim-lua/kickstart.nvim/blob/b83b2b061c1fab0e1a3a28c185345be7957e74cd/init.lua#L90

The only usecase for leader i have was ,rt to run the linter anyways.
It was such a non standard way of doing it. I am used to it and it works.
But I am trying to switch to space now to match the somewhat official
recommendation of the neovim team for the leader key.

They should know that it does not conflict with anything.
Accepting the decisions of the people that know the whole ecosystem well.
Seems like a better idea than trying to roll your own.

For example I use "c" to compile/run with my vimrc. And I love it.
But then I learned vim also uses "c" to change and its a cool vim mapping.
So I can not use that vim mapping in its intended standard way because I
rolled my own "c" compile mapping without knowing what i overwrote.